archive/recordchain_py-plainpython_impl/requirements.txt
jsonschema==2.6.0
gevent==1.2.2
merkletools==1.0.2
Flask==0.12
six==1.10.0
ACI==1.3
pycrypto==2.6.1
python_dateutil==2.6.1
text==1.0.0
